Refactor warewulfd api into its own sub-package
