+++ internal/pkg/warewulfd/api/profile.go
@@ -1,0 +1,49 @@
+package api
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/swaggest/usecase"
+	"github.com/swaggest/usecase/status"
+	"github.com/warewulf/warewulf/internal/pkg/node"
+)
+
+func getProfiles() usecase.Interactor {
+	u := usecase.NewInteractor(func(ctx context.Context, _ struct{}, output *map[string]*node.Profile) error {
+		if registry, err := node.New(); err != nil {
+			return err
+		} else {
+			*output = registry.NodeProfiles
+			return nil
+		}
+	})
+	u.SetTitle("Get node profiles")
+	u.SetDescription("Get all node profiles.")
+	u.SetTags("Profile")
+	return u
+}
+
+func getProfileByID() usecase.Interactor {
+	type getProfileByIDInput struct {
+		ID string `path:"id"`
+	}
+
+	u := usecase.NewInteractor(func(ctx context.Context, input getProfileByIDInput, output *node.Profile) error {
+		if registry, err := node.New(); err != nil {
+			return err
+		} else {
+			if profile, err := registry.GetProfile(input.ID); err != nil {
+				return status.Wrap(fmt.Errorf("profile not found: %v (%v)", input.ID, err), status.NotFound)
+			} else {
+				*output = profile
+				return nil
+			}
+		}
+	})
+	u.SetTitle("Get a node profile")
+	u.SetDescription("Get a node profile by its ID.")
+	u.SetTags("Profile")
+	u.SetExpectedErrors(status.NotFound)
+	return u
+}
